use int32_t

+++ sperl_op.h
@@ -85,7 +85,7 @@ double	64bit
 int8_t/uint8_t 8bit (1byte)
 int16_t/uint16_t 16bit (2byte)
 int32_t/uint32_t 32bit (4byte)
-int64_t/uint64_t 64bit (8byte)
+int32_t/uint32_t 64bit (8byte)
 */
 
 /* Binary operation */
@@ -101,7 +101,7 @@ struct SPerl_op {
   SPerl_OP* op_sibparent;
   union {
     char* pv;
-    int64_t iv;
+    int32_t iv;
     double nv;
   } uv;
 };
@@ -112,7 +112,7 @@ void SPerl_OpLASTSIB_set(SPerl_OP* o, SPerl_OP* parent);
 void SPerl_OpMAYBESIB_set(SPerl_OP* o, SPerl_OP* sib, SPerl_OP* parent);
 
 SPerl_OP* SPerl_newOP(int32_t type, int32_t flags, SPerl_OP *first, SPerl_OP *last);
-SPerl_OP* SPerl_op_sibling_splice(SPerl_OP* parent, SPerl_OP* start, uint64_t del_count, SPerl_OP *insert);
+SPerl_OP* SPerl_op_sibling_splice(SPerl_OP* parent, SPerl_OP* start, int32_t del_count, SPerl_OP *insert);
 SPerl_OP* SPerl_op_append_elem(SPerl_OP* first, SPerl_OP* last);
 
 void SPerl_dump_ast(SPerl_OP* op, int32_t depth);
